class SearchServer was finished and now he is working at console

+++ src/InverterIndex.cpp
@@ -3,23 +3,26 @@
 void InverterIndex::UpdateDocumentsBase(std::vector<std::string> inputDocs) {
     setlocale(LC_ALL, "English");
 
-    // ЗАПОЛНЕНИЕ docs
+    // ЗАПОЛНЕНИЕ docs ПОКА НЕ ОБНОВЛЯЕТ, А ПЕРЕЗАПИСЫВАЕТ
     docs.resize(inputDocs.size());
-    for (int i = 0; i < inputDocs.size(); i++) {
+    for (size_t i = 0; i < inputDocs.size(); i++) {
         docs[i] = inputDocs[i];
     }
     // ЗАПОЛНЕНИЕ freqDictionary
     for (size_t i = 0; i < docs.size(); i++) { // ПО ДОКУМЕНТАМ
         std::string textTemp = "";
         for (size_t k = 0; k < docs[i].length(); k++) { // ПО СИМВОЛАМ В КАЖДОЙ СТРОКЕ-ДОКУМЕНТЕ
-            if ( ((docs[i][k] >= '0' && docs[i][k] <= '9') || (docs[i][k] >= 'A' && docs[i][k] <= 'Z') || (docs[i][k] >= 'a' && docs[i][k] <= 'z')
-            || (docs[i][k] == '\'')) && (k != docs[i].length()-1)) {
+            if ( ((docs[i][k] >= '0' && docs[i][k] <= '9') || (docs[i][k] >= 'A' && docs[i][k] <= 'Z')
+            || (docs[i][k] >= 'a' && docs[i][k] <= 'z')
+            || (docs[i][k] == '\'')) && (k != docs[i].length() - 1)) {
                 textTemp += (docs[i][k]);
             } else {
-                if (textTemp == "") continue; // КРАЙНИЙ СЛУЧАЙ
-                if (k == docs[i].length()-1) {
+                if (textTemp == "") continue;
+
+                if (k == docs[i].length() - 1) {
                     textTemp += (docs[i][k]);
-                } // КРАЙНИЙ СЛУЧАЙ
+                }
+
                 auto wordInDictionary = freqDictionary.find(textTemp);
                 if (wordInDictionary == freqDictionary.end()) {
                     std::vector<Entry> tempVectorInfoWord;
@@ -33,13 +36,11 @@ void InverterIndex::UpdateDocumentsBase(std::vector<std::string> inputDocs) {
                     for (auto& it : wordInDictionary->second) {
                         if (it.doc_id == i) {
                             it.count++;
-                            //std::cout << "IN " << it.doc_id << " repeat:" << wordInDictionary->first << ", now their count is " << it.count <<std::endl;
                             idIsFound = true;
                             break;
                         }
                     }
                     if (!idIsFound) {
-                        //std::cout << "IN ANOTHER FILE: "<< wordInDictionary->first << "\n";
                         Entry newEntryInAnotherFile = {i, 1};
                         wordInDictionary->second.push_back(newEntryInAnotherFile);
                     }
@@ -57,3 +58,5 @@ std::vector<Entry> InverterIndex::GetWordCount(const std::string &word) {
     }
     return std::vector<Entry>();
 }
+
+
